some changes

+++ src/main/java/org/example/travel_agency/model/Tour.java
@@ -1,9 +1,12 @@
 package org.example.travel_agency.model;
 
 import jakarta.persistence.*;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
-
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Entity
@@ -27,7 +30,7 @@ public class Tour {
     private String type;
 
     @Column(nullable = false)
-    private double price;
+    private BigDecimal price;
 
     @Column(nullable = false)
     private LocalDate startDate;
